+++ app/javascript/vue/tasks/digitize/store/mutations/TaxonDetermination/setTaxonDetermination.js
@@ -1,3 +1,0 @@
-export default (state, value) => {
-  state.taxon_determination = value
-}