+++ app/javascript/vue/tasks/digitize/store/mutations/resetStore.js
@@ -1,4 +1,10 @@
 import { makeInitialState } from '../store.js'
+import { EVENT_TAXON_DETERMINATION_FORM_RESET } from 'constants/index.js'
+
+const resetTaxonDeterminationForm = () => {
+  const event = new CustomEvent(EVENT_TAXON_DETERMINATION_FORM_RESET)
+  document.dispatchEvent(event)
+}
 
 export default state => {
   const {
@@ -10,4 +16,6 @@ export default state => {
   state = Object.assign(state, makeInitialState())
   state.preferences = preferences
   state.project_preferences = project_preferences
+
+  resetTaxonDeterminationForm()
 }